+++ demisto_sdk/commands/common/hook_validations/conf_json.py
@@ -15,8 +15,9 @@ class ConfJsonValidator(BaseValidator):
     """
     CONF_PATH = "./Tests/conf.json"
 
-    def __init__(self, ignored_errors=None, print_as_warnings=False):
-        super().__init__(ignored_errors=ignored_errors, print_as_warnings=print_as_warnings)
+    def __init__(self, ignored_errors=None, print_as_warnings=False, suppress_print=False):
+        super().__init__(ignored_errors=ignored_errors, print_as_warnings=print_as_warnings,
+                         suppress_print=suppress_print)
         self._is_valid = True
         self.conf_data = self.load_conf_file()