+++ demisto_sdk/commands/common/tests/release_notes_test.py
@@ -52,6 +52,7 @@ def test_rn_master_diff(release_notes, expected_result, mocker):
     mocker.patch.object(ReleaseNotesValidator, '__init__', lambda a, b: None)
     ReleaseNotesValidator.ignored_errors = []
     validator = get_validator(release_notes)
+    validator.suppress_print = False
     assert validator.is_file_valid() == expected_result
 
 
@@ -158,6 +159,7 @@ def test_are_release_notes_complete(release_notes, complete_expected_result, moc
     mocker.patch.object(ReleaseNotesValidator, '__init__', lambda a, b: None)
     mocker.patch.object(StructureValidator, 'scheme_of_file_by_path', return_value='integration')
     validator = get_validator(release_notes, MODIFIED_FILES)
+    validator.suppress_print = False
     assert validator.are_release_notes_complete() == complete_expected_result
 
 
@@ -193,6 +195,7 @@ def test_are_release_notes_complete_invalid_file_type(release_notes, complete_ex
     mocker.patch.object(ReleaseNotesValidator, '__init__', lambda a, b: None)
     mocker.patch.object(StructureValidator, 'scheme_of_file_by_path', return_value='integration')
     validator = get_validator(release_notes, MODIFIED_FILES_INVALID)
+    validator.suppress_print = False
     assert validator.are_release_notes_complete() == complete_expected_result
 
 
@@ -229,6 +232,7 @@ def test_are_release_notes_complete_added(release_notes, complete_expected_resul
     mocker.patch.object(ReleaseNotesValidator, '__init__', lambda a, b: None)
     mocker.patch.object(StructureValidator, 'scheme_of_file_by_path', return_value='integration')
     validator = get_validator(release_notes, MODIFIED_FILES, ADDED_FILES)
+    validator.suppress_print = False
     assert validator.are_release_notes_complete() == complete_expected_result
 
 
@@ -271,4 +275,5 @@ def test_has_release_notes_been_filled_out(release_notes, filled_expected_result
     mocker.patch.object(ReleaseNotesValidator, '__init__', lambda a, b: None)
     mocker.patch.object(StructureValidator, 'scheme_of_file_by_path', return_value='integration')
     validator = get_validator(release_notes, MODIFIED_FILES)
+    validator.suppress_print = False
     assert validator.has_release_notes_been_filled_out() == filled_expected_result